Show signature label only when email is signed #734
In order to show the correct label in all cases we have to
PR big codes to LEAP, and we are not willing to do that at
the moment.

Besides that, Virginia was suggesting only show positive info,
for instance, not showing 'not signed' label.

Because that we decided to remove all labels but signed one.